ggml : factor all quantization code in ggml-quants
ggml-ci
